+++ cli/cli/kurtosis_gateway/server/engine_gateway/engine_gateway_service_server.go
@@ -4,19 +4,27 @@ import (
 	"context"
 	"fmt"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_gateway/connection"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_gateway/live_engine_client_supplier"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_gateway/port_utils"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/kurtosis_gateway/run/api_container_gateway"
+	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/kurtosis_core_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis-engine-api-lib/api/golang/kurtosis_engine_rpc_api_bindings"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
+	"google.golang.org/grpc"
 	"google.golang.org/protobuf/types/known/emptypb"
 	restclient "k8s.io/client-go/rest"
 	"sync"
+	"time"
 )
 
 const (
 	// API Container gateways spun up by this engine gateway run on localhost
 	localHostIpStr = "127.0.0.1"
+
+	waitForGatewayGrpcReady = true
+
+	apiContainerGatewayHealthcheckTimeout = 5 * time.Second
 )
 
 type EngineGatewayServiceServer struct {
@@ -24,7 +32,7 @@ type EngineGatewayServiceServer struct {
 	kurtosis_engine_rpc_api_bindings.UnimplementedEngineServiceServer
 
 	// Client for the engine we'll be connecting too
-	remoteEngineClient kurtosis_engine_rpc_api_bindings.EngineServiceClient
+	engineClientSupplier *live_engine_client_supplier.LiveEngineClientSupplier
 
 	// Configuration for the kubernetes cluster kurtosis is running on
 	kubernetesConfig *restclient.Config
@@ -44,12 +52,12 @@ type runningApiContainerGateway struct {
 
 // NewEngineGatewayServiceServer returns a EngineGatewayServiceServer
 // runningEngine is a kurtosis engine running a cluster that can be reached through clients configured with kubernetesConfig
-func NewEngineGatewayServiceServer(connectionProvider *connection.GatewayConnectionProvider, engineClient kurtosis_engine_rpc_api_bindings.EngineServiceClient) (resultGatewayService *EngineGatewayServiceServer, gatewayCloseFunc func()) {
+func NewEngineGatewayServiceServer(connectionProvider *connection.GatewayConnectionProvider, engineClientSupplier *live_engine_client_supplier.LiveEngineClientSupplier) (resultGatewayService *EngineGatewayServiceServer, gatewayCloseFunc func()) {
 	// We start out with no enclave api-container gateways runnings
 	runningApiContainers := map[string]*runningApiContainerGateway{}
 
 	service := &EngineGatewayServiceServer{
-		remoteEngineClient:           engineClient,
+		engineClientSupplier:         engineClientSupplier,
 		connectionProvider:           connectionProvider,
 		enclaveIdToRunningGatewayMap: runningApiContainers,
 		mutex:                        &sync.Mutex{},
@@ -65,7 +73,11 @@ func NewEngineGatewayServiceServer(connectionProvider *connection.GatewayConnect
 }
 
 func (service *EngineGatewayServiceServer) CreateEnclave(ctx context.Context, args *kurtosis_engine_rpc_api_bindings.CreateEnclaveArgs) (*kurtosis_engine_rpc_api_bindings.CreateEnclaveResponse, error) {
-	remoteEngineResponse, err := service.remoteEngineClient.CreateEnclave(ctx, args)
+	remoteEngineClient, err := service.engineClientSupplier.GetEngineClient()
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "Expected to be able to get a client for a live Kurtosis engine, instead a non nil error was returned")
+	}
+	remoteEngineResponse, err := remoteEngineClient.CreateEnclave(ctx, args)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred creating an enclave through the remote engine")
 	}
@@ -73,7 +85,7 @@ func (service *EngineGatewayServiceServer) CreateEnclave(ctx context.Context, ar
 	defer func() {
 		if cleanUpEnclave {
 			destroyEnclaveArgs := &kurtosis_engine_rpc_api_bindings.DestroyEnclaveArgs{EnclaveId: args.GetEnclaveId()}
-			if _, err := service.remoteEngineClient.DestroyEnclave(ctx, destroyEnclaveArgs); err != nil {
+			if _, err := remoteEngineClient.DestroyEnclave(ctx, destroyEnclaveArgs); err != nil {
 				logrus.Error("Launching the Enclave gateway failed, expected to be able to cleanup the created enclave, but an error occurred calling the backend to destroy the enclave we created:")
 				fmt.Fprintln(logrus.StandardLogger().Out, err)
 				logrus.Errorf("ACTION REQUIRED: You'll need to manually kill the enclave with id '%v'", args.GetEnclaveId())
@@ -108,7 +120,11 @@ func (service *EngineGatewayServiceServer) CreateEnclave(ctx context.Context, ar
 }
 
 func (service *EngineGatewayServiceServer) GetEnclaves(ctx context.Context, in *emptypb.Empty) (*kurtosis_engine_rpc_api_bindings.GetEnclavesResponse, error) {
-	remoteEngineResponse, err := service.remoteEngineClient.GetEnclaves(ctx, in)
+	remoteEngineClient, err := service.engineClientSupplier.GetEngineClient()
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "Expected to be able to get a client for a live Kurtosis engine, instead a non nil error was returned")
+	}
+	remoteEngineResponse, err := remoteEngineClient.GetEnclaves(ctx, in)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting info for enclaves from the remote engine")
 	}
@@ -117,14 +133,14 @@ func (service *EngineGatewayServiceServer) GetEnclaves(ctx context.Context, in *
 	for enclaveId, enclaveInfo := range responseEnclaves {
 		var runningApiContainerGateway *runningApiContainerGateway
 		runningApiContainerGateway, isRunning := service.enclaveIdToRunningGatewayMap[enclaveId]
-		defer func() {
-			if cleanUpRunningGateways {
-				service.idempotentKillRunningGatewayForEnclaveId(enclaveId)
-			}
-		}()
 		// If the gateway isn't running, start it
 		if !isRunning {
 			runningApiContainerGateway, err = service.startRunningGatewayForEnclave(enclaveInfo)
+			defer func() {
+				if cleanUpRunningGateways {
+					service.idempotentKillRunningGatewayForEnclaveId(enclaveId)
+				}
+			}()
 			if err != nil {
 				return nil, stacktrace.Propagate(err, "Expected to be able to start a local gateway for enclave '%v', instead a non-nil error was returned", enclaveId)
 			}
@@ -137,7 +153,11 @@ func (service *EngineGatewayServiceServer) GetEnclaves(ctx context.Context, in *
 }
 
 func (service *EngineGatewayServiceServer) StopEnclave(ctx context.Context, args *kurtosis_engine_rpc_api_bindings.StopEnclaveArgs) (*emptypb.Empty, error) {
-	if _, err := service.remoteEngineClient.StopEnclave(ctx, args); err != nil {
+	remoteEngineClient, err := service.engineClientSupplier.GetEngineClient()
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "Expected to be able to get a client for a live Kurtosis engine, instead a non nil error was returned")
+	}
+	if _, err := remoteEngineClient.StopEnclave(ctx, args); err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred calling remote engine to stop enclave '%v'", args.EnclaveId)
 	}
 
@@ -145,7 +165,11 @@ func (service *EngineGatewayServiceServer) StopEnclave(ctx context.Context, args
 }
 
 func (service *EngineGatewayServiceServer) DestroyEnclave(ctx context.Context, args *kurtosis_engine_rpc_api_bindings.DestroyEnclaveArgs) (*emptypb.Empty, error) {
-	if _, err := service.remoteEngineClient.DestroyEnclave(ctx, args); err != nil {
+	remoteEngineClient, err := service.engineClientSupplier.GetEngineClient()
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "Expected to be able to get a client for a live Kurtosis engine, instead a non nil error was returned")
+	}
+	if _, err := remoteEngineClient.DestroyEnclave(ctx, args); err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred calling remote engine to destroy enclave with ID '%v':", args.EnclaveId)
 	}
 	// Kill the running api container gateway for this enclave
@@ -156,7 +180,11 @@ func (service *EngineGatewayServiceServer) DestroyEnclave(ctx context.Context, a
 }
 
 func (service *EngineGatewayServiceServer) Clean(ctx context.Context, args *kurtosis_engine_rpc_api_bindings.CleanArgs) (*kurtosis_engine_rpc_api_bindings.CleanResponse, error) {
-	remoteEngineResponse, err := service.remoteEngineClient.Clean(ctx, args)
+	remoteEngineClient, err := service.engineClientSupplier.GetEngineClient()
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "Expected to be able to get a client for a live Kurtosis engine, instead a non nil error was returned")
+	}
+	remoteEngineResponse, err := remoteEngineClient.Clean(ctx, args)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred while cleaning enclaves")
 	}
@@ -168,7 +196,11 @@ func (service *EngineGatewayServiceServer) Clean(ctx context.Context, args *kurt
 }
 
 func (service *EngineGatewayServiceServer) GetEngineInfo(ctx context.Context, emptyArgs *emptypb.Empty) (*kurtosis_engine_rpc_api_bindings.GetEngineInfoResponse, error) {
-	remoteEngineResponse, err := service.remoteEngineClient.GetEngineInfo(ctx, emptyArgs)
+	remoteEngineClient, err := service.engineClientSupplier.GetEngineClient()
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "Expected to be able to get a client for a live Kurtosis engine, instead a non nil error was returned")
+	}
+	remoteEngineResponse, err := remoteEngineClient.GetEngineInfo(ctx, emptyArgs)
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting engine info through the remote engine")
 	}
@@ -186,7 +218,7 @@ func (service *EngineGatewayServiceServer) startRunningGatewayForEnclave(enclave
 		return nil, stacktrace.Propagate(err, "Expected to be able to get a free, open TCP port on host, instead a non-nil error was returned")
 	}
 	// Channel for messages to stop the running server
-	gatewayServerStopper := make(chan interface{}, 1)
+	gatewayStopChannel := make(chan interface{}, 1)
 
 	// Info for how to connect to the api container through the gateway running on host machine
 	apiContainerHostMachineInfo := &kurtosis_engine_rpc_api_bindings.EnclaveAPIContainerHostMachineInfo{
@@ -200,10 +232,12 @@ func (service *EngineGatewayServiceServer) startRunningGatewayForEnclave(enclave
 	// Stop the running gateway
 	gatewayStopFunc := func() {
 		// Send message to stop the gateway server
-		gatewayServerStopper <- nil
+		gatewayStopChannel <- nil
 	}
+	// TODO: Modify MinimalGrpcServer.RunUntilStopped to take in a `ReadyChannel` to communicate when a GRPC server is ready to serve
+	// Currently, we have to make a health check request to verify that the API container gateway is ready
 	go func() {
-		if err := api_container_gateway.RunApiContainerGatewayUntilStopped(service.connectionProvider, enclaveInfo, gatewayPortSpec.GetNumber(), gatewayServerStopper); err != nil {
+		if err := api_container_gateway.RunApiContainerGatewayUntilStopped(service.connectionProvider, enclaveInfo, gatewayPortSpec.GetNumber(), gatewayStopChannel); err != nil {
 			logrus.Warnf("Expected to run api container gateway until stopped, but the server exited prematurely with a non-nil error: '%v'", err)
 		}
 	}()
@@ -213,12 +247,15 @@ func (service *EngineGatewayServiceServer) startRunningGatewayForEnclave(enclave
 			gatewayStopFunc()
 		}
 	}()
+	// Need to wait for the GRPC server spun up in the goFunc to be ready
+	if err := waitForGatewayReady(apiContainerHostMachineInfo); err != nil {
+		logrus.Errorf("Expected Gateway to be reachable, instead an error was returned:\n%v", err)
+	}
 
 	runningGatewayInfo := &runningApiContainerGateway{
 		closeFunc:       gatewayStopFunc,
 		hostMachineInfo: apiContainerHostMachineInfo,
 	}
-
 	// Store information about our running gateway
 	service.enclaveIdToRunningGatewayMap[enclaveId] = runningGatewayInfo
 	cleanUpMapEntry := true
@@ -233,6 +270,31 @@ func (service *EngineGatewayServiceServer) startRunningGatewayForEnclave(enclave
 	return runningGatewayInfo, nil
 }
 
+// Calls `GetModules` and waits for the gateway to be ready
+func waitForGatewayReady(apiContainerHostMachineInfo *kurtosis_engine_rpc_api_bindings.EnclaveAPIContainerHostMachineInfo) error {
+	backgroundCtx := context.Background()
+	gatewayAddress := fmt.Sprintf("%v:%v", apiContainerHostMachineInfo.IpOnHostMachine, apiContainerHostMachineInfo.GrpcPortOnHostMachine)
+
+	conn, err := grpc.Dial(gatewayAddress, grpc.WithInsecure())
+	if err != nil {
+		return stacktrace.Propagate(err, "Expected to be dial in to API container running at address '%v', instead a non-nil error was returned", gatewayAddress)
+	}
+	apiContainerClient := kurtosis_core_rpc_api_bindings.NewApiContainerServiceClient(conn)
+
+	// The GRPC Server for our `API Container Gateway` is spun up in a gofunc
+	// We call `GetModules` with `WaitForReady` to wait for the server to finish setting up
+	// Modifying
+	ctxWithTimeout, cancelFunc := context.WithTimeout(backgroundCtx, apiContainerGatewayHealthcheckTimeout)
+	defer cancelFunc()
+	getModulesHealthCheckParams := &kurtosis_core_rpc_api_bindings.GetModulesArgs{Ids: nil}
+	_, err = apiContainerClient.GetModules(ctxWithTimeout, getModulesHealthCheckParams, grpc.WaitForReady(waitForGatewayGrpcReady))
+	if err != nil {
+		return stacktrace.Propagate(err, "Expected to be to call `GetModules` and wait for server to be ready, instead a non-nil error was returned")
+	}
+
+	return nil
+}
+
 func (service *EngineGatewayServiceServer) idempotentKillRunningGatewayForEnclaveId(enclaveId string) {
 	service.mutex.Lock()
 	defer service.mutex.Unlock()
@@ -244,6 +306,8 @@ func (service *EngineGatewayServiceServer) idempotentKillRunningGatewayForEnclav
 
 	// Close up the connections
 	runningGateway.closeFunc()
+
+	logrus.Infof("Stopped running Gateway for enclave '%v' on port '%v'", enclaveId, runningGateway.hostMachineInfo.GrpcPortOnHostMachine)
 	// delete the entry for the enclave
 	delete(service.enclaveIdToRunningGatewayMap, enclaveId)
 }